server/requirements.txt
openenv-core>=0.1.0
pydantic>=2.0
uvicorn>=0.27.0
fastapi>=0.109.0
